requirements.txt
streamlit
pandas
plotly
pybit
PyYAML
python-dotenv
requests
